Split up app controller (routes) and main server connection logic
